# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LensFlow AI - Real estate video pipeline API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companion/subscribe/checkout:
    post:
      operationId: createCompanionCheckout
      tags: [companion]
      summary: Create a Stripe Checkout session for a Glimr subscription tier (Spark/Flame)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionSubscribeCheckoutBody"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionSubscribeCheckoutResponse"
        "400":
          description: Invalid input
        "502":
          description: Stripe checkout session creation failed
  /companion/subscribe/verify:
    post:
      operationId: verifyCompanionCheckout
      tags: [companion]
      summary: Verify a completed Stripe Checkout session and activate the subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionSubscribeVerifyBody"
      responses:
        "200":
          description: Subscriber activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionSubscribeVerifyResponse"
        "400":
          description: Checkout session is not a completed subscription
  /companion/subscribe/status:
    post:
      operationId: getCompanionSubscribeStatus
      tags: [companion]
      summary: Get the current subscription tier and voice-message usage for an email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionSubscribeStatusBody"
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionSubscribeStatusResponse"
        "400":
          description: Invalid input
  /companion/subscribe/portal:
    post:
      operationId: createCompanionPortal
      tags: [companion]
      summary: Create a Stripe billing portal session for a subscriber to manage/cancel their plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanionSubscribePortalBody"
      responses:
        "200":
          description: Billing portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanionSubscribePortalResponse"
        "404":
          description: No billing account found for that email
        "502":
          description: Stripe portal session creation failed
components:
...
  schemas:
...
    CompanionSubscribeCheckoutBody:
      type: object
      required: [tier]
      properties:
        tier:
          type: string
          enum: [spark, flame]
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
          description: Optional known email — prefills and locks the Stripe customer instead of asking Checkout for one.
    CompanionSubscribeCheckoutResponse:
      type: object
      required: [checkoutUrl]
      properties:
        checkoutUrl:
          type: string
    CompanionSubscribeVerifyBody:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
          minLength: 1
    CompanionSubscribeVerifyResponse:
      type: object
      required: [email, tier]
      properties:
        email:
          type: string
        tier:
          type: string
          enum: [free, spark, flame]
    CompanionSubscribeStatusBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
    CompanionSubscribeStatusResponse:
      type: object
      required: [tier, active, voiceLimit, voiceRemaining]
      properties:
        tier:
          type: string
          enum: [free, spark, flame]
        active:
          type: boolean
        voiceLimit:
          type: ["integer", "null"]
        voiceRemaining:
          type: ["integer", "null"]
    CompanionSubscribePortalBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
    CompanionSubscribePortalResponse:
      type: object
      required: [portalUrl]
      properties:
        portalUrl:
          type: string
...